backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
openaq==2.0.0
